requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-flat-theme==0.9.5
django-toolbelt==0.0.1
djangorestframework==3.1.3
gunicorn==19.3.0
Pillow==2.9.0
psycopg2==2.6.1
static3==0.6.1
Unipath==1.1
